Update backend and frontend code
In the backend, the MySQL database connection settings have been changed in the index.php file. The plugin-api-version has been downgraded to 2.3.0 in the composer.lock file. A new route has been added to the index.php file to retrieve all alumnos from the database. A new route has also been added to insert new alumnos into the database.

In the frontend, a new button has been added to each alumno in the list to delete the alumno. A new method has been added to the Vue instance to handle the deletion of alumnos. The obtenerAlumnos method has been updated to use the new route added in the backend. The crearAlumno method has been updated to use the new route added in the backend. The nuevoAlumno data property has been reset after creating a new alumno.